Implement decoding and encoding of configuration.
A YAML config file can be provided to the `parse` command, which will
convert it into an opinionated AWS CloudFormation template, creating a
load balancer, auto scaling group, security group, and launch
configuration based on the configuration.

The generated template assumes a rolling update policy for the
autoscaling group, an ELB based health check, and requires a VPC to
provision resources in. It is currently opinionated with regard to
timeouts and intervals; this may change in the future.

Subnets can be provided for both the load balancer and auto scaling
group, allowing private instances and a public load balancer.

The generated template is output to stdout.